First draft of the DB functions
The next version of the WMS Inspector will include an SQLite database to store
user data locally. WMSInspector.DB will contain all the related methods,
including the ones related with upgrading the database schema in future
versions of the add-on. Right now, every time the extension is loaded, it
checks if the database exists. If not, it copies the empty database located in
the defaults directory to the profile folder. If the database exists but it
has an old schema, it will be upgraded to the latest version.
TODO This check should be probably only needed when the extension is first
installed or updated