Search functionality in Library
The WMSInspector.libraryQuery allows to perform queries in the services
database. It accepts a WMSInspector.libraryQueryParams object with the
required parameters and a callback function to be called after the query is
performed.
It is important to notice that the classes and methods used are only
compatible with versions of Firefox greater or equal to 3.5. This will be in
fact the minimum required version of Firefox for the next release (0.1.5).