Open GetCapabilities response from Library
Users can access both HTML reports or original XML files from GetCapabilities
responses of the services stored in the Library. This is done via the context
menu. As currently HTML reports are only available to WMS services, if another
type of service is selected, the context menu entry is disabled.

The method WMSInspector.Overlay.getGetCapabilitiesURL() has been created to
avoid duplication of code in different parts of the application.