Import and export functions in the Library
Users can now export and import lists of services from and to the Library. The
files used are CSV (The separator can be configured via preferences, but it is
not shown in the options dialog). Users can export all services in the DB or
just the results shown in the services list. As import operations can be both
time and resource consuming, new functionality to run separated threads has been
implemented. Processes running in a background thread must never try to access
elements from the UI or any other interface that is not thread safe.
WMSInspector.IO now has the read() and readLineByLine() methods to read the
contents of a local file.