New XPCOM component and JS module
Some major changes in the general structure of the code have been
performed.

An XPCOM component (WMSInspectorService) will handle the thread
related code regarding the import of CSV files to the library, as
threads can not be performed in the window scope. If possible, all DB
related functions (add/update/delete/query services) should end up in
this component.

A new module has been added to hold the definitions of classes used in
different parts of the code (like the former libraryService). They all
will share the WMSInspectorClasses namespace. If possible, all the
classes scattered through the code should be placed in here.

The exception handling in the Library functions has been improved.