Changes in the WMSInspector namespace
In to order to simplify the code and optimize the way in which scripts are
loaded, this changeset includes the following points:
-  The WMSInspector namespace is only used in the browser overlay and opened
   windows and dialogs don't use the WMSInspector namespace
-  The WMSInspectorClasses namespace has been deleted, and the Classes object
   will be loaded in the WMSInspector when loading it from the overlay or in
   the dialog scope if loaded from a dialog
DB, IO , Utils, Requests and all other possible functions will be migrated to
Javascript code modules. The scripts that need some of this functions will
load them in the appropiate scope.