Move Utils to a Javascript Code Module
Modules allow the reusing of code between different privileged environments
like window scripts or XPCOM components. Modules are cached when loaded and
subsequent imports will share the same objects and methods in the different
scopes.
From the scope of a module, propoerties like 'window' and 'document' can not
be accessed, so the equivalent XPCOM services have been used. This makes
unnecessary using stringBundle elements in XUL windows and dialogs.