Move service related functions to component
The functions for adding, updating and deleting services have been moved
to the WMS Inspector XPCOM component. This will allow to reuse them in
different parts of the application without needing to load all the
Library functions, as well as allowing to run them threaded.

The public methods are exposed by the WMSInspectorService instances,
but the actual work is performed by the methods of
WMSInspectorServicePrivate, which also includes the thread methods.