Create a getTags function in component
The getTags function returns an array of all the current tags in the
DB. It is used by the Library and the AddServiceDialog.
With this change, all DB related functions have been moved to the
component and the modules. Overlay and AddServiceDialog no longer need
to hold a reference to the Library object.