Update startup version management
Firefox 4.0 deprecated the extension manager service in favor of an
AddonManager object. The related functions are asynchronous, so the
functions called when the overlat is loaded have been moved to an
onReady() merthod. This is actually one of the best practices that are
encouraged for Firefox 4.0. The old code has been preserved for
backwards compatibility, but should be removed when the support for
Firefox 3.6 is dropped.