Add new logging system
To support the development and debugging, a new logging system has
been implemented. It supports both console and file output. Both are
enabled by default, and can be disabled using via the preferences.
The file output is based on the log4moz script, and the file is
located on the wmsinspector profile dir. Once loaded the Log object,
the following functions can be called: error(), warn(), info() and
debug(). If the parameter sent is an Error or an nsIException object,
the relevant properties wil be logged. If it is an object, it will be
serialized. Otherwise, it will be shown as a plain message.