Open application in a new window (Issue #28)
Users can now detach the application from the main window and reatach
it. The XUL overlays have been redefined, and now there is a common
one (wmsinspectorOverlay.xul), one for the panel (browserOverlay.xul)
and one for the window (wmsinspectorWindow.xul). When opening a new
window, it has a new scope, so to keep the state (services loaded,
expanded leafs, group mode, etc.), we pass a state object to try to
restore it.