New logos
The next release will include a brand new logo and icons. The main
button will be still be located in the addons bar, but users will be
able to place it in any of the toolbars of the browser (also the
library button). Different sizes and colors have been provided for
different platforms, following the instructions detailed in this
post:

http://blog.mozilla.com/addons/2010/12/02/toolbar-buttons-in-firefox-4/

Suitable icons for the new addons manager have been also addded.