Support for packed extension in Firefox 4
On Firefox 4, the extensions are not extracted in the profile
directory. Instead of this, the XPI file is copied directly, without
extracting its contents. As a consequence, the default DB file can
not be directly copied to the profile dir on the first run. Instead,
the file must be read and write with the new IO.saveURItoFile()
function (To get the URI, we use addon.getResourceURI from the
AddonManager interface.

As a result of this, setting up the DB is now an async operation, so
callers of DB.checkDB or DB.restoreDB can provide a callback function
(i.e. onReady in Overlay and Library).